Style and implement several components
CountryDetails was implemented to render the
appropriate content when a card is selected from
the Home page. Filters and Header were styled,
adjusting some shadows and padding. App and
CountryCard were altered slightly to change the
way the react-router handles the routes and base
name, to properly render the CountryDetails based
on the selected country name.